requirements.txt
mdtraj
pyyaml>=6.0
pytorch-lightning[extra]
tqdm
ase
e3nn
numpy
scipy
scikit-learn
pytest
